+++ R/serp.fit.R
@@ -2,7 +2,7 @@
 
 serpfit <- function(x, y, wt, yMtx, link, slope, reverse, control,
                     Terms, lambda, lambdaGrid, gridType, tuneMethod,
-                    globalEff, cvMetric, mslope, linkf, nL, obs,
+                    globalEff, cvMetric, mslope, nL, obs,
                     vnull, nvar, m)
 {
   if (slope == 'penalize' && !is.null(globalEff))
@@ -24,6 +24,22 @@ serpfit <- function(x, y, wt, yMtx, link, slope, reverse, control,
         stop("response not allowed in 'globalEff'")
     } else stop("'globalEff' is unspecified")
   }
+  link_reversed <- FALSE
+  initial_link <- link
+
+  if (link == "cloglog" & reverse == TRUE){
+    link_reversed <- TRUE
+    link <- "loglog"
+    initial_link <- "cloglog"
+  }
+  if(initial_link == "loglog"){
+    if (link == "loglog" & reverse == TRUE){
+      link_reversed <- TRUE
+      link <- "cloglog"
+      initial_link <- "loglog"
+    }
+  }
+  linkf <- lnkfun(link)
   xlst <- formxL(x, nL, slope, globalEff, m, vnull)
   yFreq <- colSums(yMtx)/obs
   xMat <- do.call(rbind, xlst)
@@ -143,44 +159,38 @@ serpfit <- function(x, y, wt, yMtx, link, slope, reverse, control,
   misc <- list(colnames.x = colnames.x, colnames.xMat = nmsv, npar = npar,
                variable.null = vnull, convg.no = res$conv,  no.var = nvar)
   delta <- res$coef
-  fv <- res$exact.pr
-  if (reverse) {
-    reverse.par <- reverse.fun(delta, slope, globalEff,
-                               m, mslope, fv, nL, Terms, misc)
-    delta <- reverse.par[[1L]]
-    fv <- reverse.par[[2L]]
-  }
-  fv <- as.data.frame(fv)
+  if (reverse) delta <- -1L * delta
+  fv <- as.data.frame(res$exact.pr)
   colnames(fv) <- levels(y)
+  nfinite <- "non-finite log-likelihood persists, "
   if (!is.null(globalEff)) ans$globalEff <- globalEff
   if (slope == "penalize"){
     if (is.na(res$loglik)){
       if (tuneMethod=="user")
-        warning("non-finite log-likelihood persists, ",
-                "try larger values of lambda.")
+        warning(nfinite,"apply a stronger lambda value or a different ",
+                "tuning method")
       if (tuneMethod=="deviance" || tuneMethod=="cv"){
         if (!is.null(lambdaGrid))
-          warning("non-finite log-likelihood persists, ",
-                  "try increasing lambdaGrid upper limit.")
+          warning(nfinite,"increase lambdaGrid upper limit or apply a different ",
+                  "tuning method")
         else
-          warning("non-finite log-likelihood persists, ",
-                  "try other tuning methods.")
+          warning(nfinite,"try a different tuning method.")
       }
     }
     if (tuneMethod == "deviance" || tuneMethod == "cv"){
       if (tuneMethod == "cv"){
-        ans$testError <- minL$objective
-        ans$trainError <- errorMetrics(y, fv, type = cvMetric)
+        ans$testError <- if (is.na(res$loglik)) NA else minL$objective
       } else ans$value <- as.numeric(minL$objective)
-      if (gridType == "discrete") {
+      if (gridType == "discrete"){
         misc$gridType <- gridType
-        misc$grid.range <- c(0L, control$maxpen)
+        misc$grid.range <- if(!is.null(lambdaGrid))
+          c(0L, max(lambdaGrid)) else c(0L, control$maxpen)
       } else {
-        misc$grid.range <- range(lambdaGrid)
+        misc$grid.range <- c(0L, control$maxpen)
         ans$gridType <- gridType
       }
     }
-    ans$lambda <- lam
+    ans$lambda <- if (is.na(res$loglik)) NA else lam
     if (tuneMethod == "finite") ans$value <- res$loglik
   }
   if (!vnull){
@@ -197,6 +207,7 @@ serpfit <- function(x, y, wt, yMtx, link, slope, reverse, control,
     dimnames(hes) <- list(new.name, new.name)
     names(gra) <- new.name
   }
+  if (link_reversed) link <- initial_link
   delta <- as.numeric(delta)
   sl <- if (partial.names) "partial" else slope
   delta <- est.names(delta, slope = sl, globalEff, colnames.x, x, m,
@@ -293,6 +304,9 @@ serp.fit <- function(lambda, x, y, wt, startval, xlst,
     }
     converged <- abs.conv
     half.iter <- 0L
+    delta.old <- delta
+    obj.old <- obj
+    fvalues.old <- fvalues
     while (obj < objOld && (conv == 2L)) {
       delta <- (delta + deltaOld) * 0.5
       fvalues <- prlg(delta, xMat, obs, yMtx, penx, linkf, control, wt)
@@ -309,11 +323,11 @@ serp.fit <- function(lambda, x, y, wt, startval, xlst,
         break
       }
     }
-    if(conv == 3L) break
-    Improved <- objOld <= obj
-    if (!Improved) {
-      delta <- deltaOld
-      loglik <- objOld
+    NotImproved <- objOld <= obj
+    if (NotImproved) {
+      delta <- delta.old
+      loglik <- obj.old
+      fvalues <- fvalues.old
     }
   }
   if (maxits > 1L && iter >= maxits){