+++ R/serp.misc.R
@@ -120,6 +120,8 @@ prlg <- function(delta, xMat, obs, yMtx, penx, linkf, control, wt)
     cbind(eta[, -ncol(eta)] - eta[, -1L], eta[,dim(eta)[2L]] )
   np <- any(pr < 0)
   pr[pr <= control$minP] <- control$minP
+  pr[pr > 1-control$minP] <- 1- control$minP
+
   logpr <- suppressWarnings(log(pr))
   logL <- sum(wt * yMtx * logpr) + penx$logpen
   exact.logpr <- suppressWarnings(log(exact.pr))
@@ -140,7 +142,9 @@ ScoreInfo <- function(x, y, pr, wt, nL, yMtx, xlst, penx, linkf)
     wts <- rowSums(yMtx) else wts <- attr(yMtx, "wts")
   pw <- wts/um$pr
   ff <- wts/um$lp
-  q <- lapply(split(um$etaMat, seq_len(nrow(um$etaMat))), etfun, nL, linkf)
+  eta <- um$etaMat
+  split.eta <- split(eta, seq_len(nrow(eta)))
+  q <- lapply(split.eta, etfun, nL, linkf)
   invmat <- lapply(split(pw, seq_len(nrow(pw))), diag)
   Infx <- mapply(
     function(invt, ff, x, q, wt){
@@ -176,9 +180,17 @@ checkArg <- function (mcall, scall, argnames)
 
 startv <- function(link, linkf, globalEff, x, yFreq, xMat, nL, nvar, slope)
 {
-  ut <- linkf$qfun(cumsum(yFreq[-nL]))
+  if (link == "logit")
+    ut <- linkf$qfun(cumsum(yFreq[-nL]))
   if (link == "cauchit")
     ut <- tan(pi*(cumsum(yFreq[-nL]) - 0.5))
+  if (link == "cloglog")
+    ut <- log(-log(1 - cumsum(yFreq[-nL])))
+  if (link == "loglog")
+    ut <- -log(-log(cumsum(yFreq[-nL])))
+  if (link == "probit")
+    ut <- qnorm(cumsum(yFreq[-nL]))
+
   if (slope == "parallel")
     ibeta <- c(ut, rep(0, nvar))
   if (slope == "partial")
@@ -278,8 +290,6 @@ cvfun <- function(lambda, x, y, nrFold, linkf, link, m, slope, globalEff,
   x <- x[Indx, ]
   y <- y[Indx]
   yMtx <- yMtx[Indx, ]
-  if (reverse) yMtx <- yMtx[ ,sort(seq_len(ncol(yMtx)),
-                                   decreasing = TRUE)]
   folds  <- cut(seq_len(obs), breaks = nrFold,
                 labels = FALSE)
   err  <- matrix(NA, nrow = nrFold, ncol = 1)
@@ -311,7 +321,7 @@ cvfun <- function(lambda, x, y, nrFold, linkf, link, m, slope, globalEff,
       obs <- length(ts.y)
       ts.xMat <- do.call(rbind, ts.xlst)
       pr <- prlg(est, ts.xMat, obs, yMtx = NULL, penx = NULL, linkf,
-                 control = NULL, wt = NULL)$pr
+                 control = NULL, wt = NULL)$exact.pr
       lv <- levels(ts.y)
       dplv <- levels(droplevels(ts.y))
       if (length(dplv) != length(lv)){
@@ -321,10 +331,12 @@ cvfun <- function(lambda, x, y, nrFold, linkf, link, m, slope, globalEff,
         pr <- pr[, which(dplv %in% lv), drop = FALSE]
       }
       pm <- errorMetrics(ts.y, pr, type = cverror)
+
     } else pm <- NA
     err[r, ] <- pm
   }
-  sum(na.omit(err))/length(err)
+  res <- na.omit(err)
+  sum(res)/length(res)
 }
 
 dvfun <- function(lambda, globalEff, x, y, startval, xlst, xMat, yMtx,
@@ -337,15 +349,18 @@ dvfun <- function(lambda, globalEff, x, y, startval, xlst, xMat, yMtx,
                    vnull,control, slope, globalEff, m,
                    mslope, tuneMethod, Terms, xtrace=FALSE);},
     error=function(e) {rr <- NA})
-  logLik <- if(tuneMethod == "finite") rr$exact.logL else rr$logL
-  tryCatch({
-    rd <- -2*as.numeric(logLik);}, error=function(e) {rd <- NA})
+  if(is.list(rr)){
+    logLik <- if(tuneMethod == "finite") rr$exact.logL else rr$logL
+    tryCatch({rd <- -2*as.numeric(logLik);}, error=function(e) {rd <- NA})
+  } else {rd <- Inf}
+
   return(rd)
 }
 
 reverse.fun <- function(delta, slope, globalEff, m, mslope, fv, nL,
                         Terms, misc)
 {
+  .Deprecated(NULL)
   if (mslope == 'penalize' && slope != 'parallel' &&
       !is.null(globalEff)) slope <- "partial"
   delta <- c(delta)
@@ -451,15 +466,11 @@ penalty.print <- function(object, max.tun)
     h2 <- round(object$lambda, 5L)
   }
   if (!tun == 'user'){
-    if(tun == "cv"){
-      h0 <- round(as.numeric(object$trainError), 6L)
+    if(tun == "cv")
       h1 <- round(as.numeric(object$testError), 6L)
-    } else {
-      h0 <- NULL
+    else
       h1 <- round(as.numeric(object$value), 6L)
-    }
     h2 <- round(as.numeric(object$lambda), 5L)
-    if (is.na(object$logLik)) h0 <- h1 <- h2 <- NA
   }
   cat("\nRegularization Info:")
   if (!length(colnames(object$model)[-1L]) ==
@@ -468,19 +479,32 @@ penalty.print <- function(object, max.tun)
     cat("\ntuneMethod:","" , tun)
     if (object$tuneMethod == "cv"){
       cat("\ncvMetric:",  "  ", object$cvMetric)
-      cat("\ntrainError:","" , h0)
       cat("\ntestError:"," " , h1)
-    } else cat("\nvalue:","     " , h1)
+    }
+    if (object$tuneMethod == "finite"){
+      if (is.na(h1)) cat("\nvalue:","     " , h1)
+      else cat("\nvalue:","    " , h1)
+    }
+    if (object$tuneMethod == "deviance" || object$tuneMethod == "user")
+      cat("\nvalue:","     " , h1)
     if (object$tuneMethod == "cv" || object$tuneMethod == "deviance"){
-      if (!is.na(h2) && length(object$lambdaGrid) > 1L &&
-          round(h2) >= round(max(object$misc$grid.range))){
-        cat("\nlambda:","    " , h2,'!')
+      if (is.null(object$gridType)){
+        len.lamG <- length(object$lambdaGrid)
+        max.lam <- round(max(object$lambdaGrid))
+      } else {
+        len.lamG <- 2L
+        max.lam <- round(object$control$maxpen)
+      }
+      if (!is.na(h2) && len.lamG > 1L &&
+          round(h2) >= max.lam){
+        cat("\nlambda:","    " , h2,'*')
         max.tun <- TRUE
       } else cat("\nlambda:","    " , h2)
     } else cat("\nlambda:","    " , h2)
   } else cat("\nno subject-specific effects found, ",
              "parallel slope(s) returned\n")
   cat("\n")
+  return(max.tun)
 }
 
 est.names <- function(coefs, slope, globalEff, colnames.x, x, m, npar,